+++ zig/http.go
@@ -1,6 +1,0 @@
-package zig
-
-type HttpServer interface {
-	Start(app *App)
-	Stop() error
-}