+++ zig/metrics.go
@@ -1,7 +1,0 @@
-package zig
-
-type MetricPublisher interface {
-	Start(app *App) error
-	Stop() error
-	IncCounter(metricName string, value int, arguments map[string]string) error
-}