+++ zig/util.go
@@ -1,11 +1,13 @@
 package zig
 
-func pipeHandlers(funcs ...Middleware) func(handlerFunc HandlerFunc) HandlerFunc {
+func PipeHandlers(funcs ...MiddlewareFunc) func(handlerFunc HandlerFunc) HandlerFunc {
 	return func(origHandler HandlerFunc) HandlerFunc {
 		var handlerResult HandlerFunc
 		return func(messageEvent MessageEvent, app *App) ProcessStatus {
-			for index, f := range funcs {
-				if index == 0 {
+			last := len(funcs) - 1
+			for i := last; i >= 0; i-- {
+				f := funcs[i]
+				if i == last {
 					handlerResult = f(origHandler)
 				} else {
 					handlerResult = f(handlerResult)