+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.5] - 2020-10-16
+### Added
+- Add tests for util
+### Changes
+- Fixes middleware execution order
+- Fixes log formatting
+- Disables colored output for logs
+
+## [v0.1.4] - 2020-10-15
+### Added
+- Adds function to DefaultHTTPServer to attach routes
+- Adds ping endpoint to DefaultHTTPServer
+
+## [v0.1.3] - 2020-10-14
+### Changes
+- Retrier interface Start method returns a channel to wait on
+- Rabbit Retrier's consumer polling is moved into Start method
+
 ## [v0.1.1] - 2020-10-12
 ### Changes
 - Fixes go mod module path